继承openfeign

+++ shop-service-pay/shop-service-pay-biz/src/main/java/com/shop/pay/core/client/dto/submit/PayOrderSubmitReqDTO.java
@@ -2,6 +2,8 @@ package com.shop.pay.core.client.dto.submit;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PayOrderSubmitReqDTO {
 
@@ -11,7 +13,7 @@ public class PayOrderSubmitReqDTO {
 
     private String outTradeNo;
 
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
 
    private String subject;
